Extract Motor control code from StrokeEngine and inject a MotorInterface (WIP)
